+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 - `cluster template` supports a generic way to template `CAPI` based clusters where all the input parameters are given as `values.yaml` from the corresponding `cluster` and `default-apps` chart.
 - CAPZ: removed unmaintained `CAPZ` implementation and switched to the generic templating implementation.
 
+### Changed
+
+- `kubectl gs template cluster` for Cluster API provider AWS has been adapted to work with the values schema of cluster-aws v0.28.0.
+
 ## [2.33.0] - 2023-03-08
 
 ### Added